mfd: axp20x: Improve interrupt order for POK handling during sleep
When the individual interrupt is not handled immediately and the
user manages to press and release the POK before interrupt handler
fires, the interrupts will get sent up the stack in the order defined
in the *regmap_irqs array. For POK, this means that release event
will get sent before the press event, and in that case the key
may get stuck in pressed mode even when the user is not holding
the key anymore.

Of course if interrupt handler doesn't run for multiple press/release
cycles, we may not know the physical state of the button, because the
flags are not getting cleared, but it's preferable to err on released
state, instead of pressed state, because that causes fewer issues
for the user. If we err on pressed state (current behavior), userspace
may simply decide to power off the users's device due to "long press"
of the power button, if the users resumes the device by pressing
and releasing the power key quickly enough, before interrupt handler
runs.

Signed-off-by: Ondrej Jirman <megi@xff.cz>