arm64: dts: rk3399-pinephone-pro: Add mount-matrix for magnetometer
After experimenting with the magnetometer on the Pinephone Pro I found out the
mount-matrix currently reported by the AF8133J driver is just the identity
matrix (no rotation), which is wrong. It would seem to me that there is
currently no `mount-matrix` info and that's why the driver reports a trivial
answer. As far as I can tell from my tests, a 90 degree clockwise rotation in
the XY plane would align it properly.

Signed-of-by: Leonardo G. Trombetta <lgtrombetta@gmx.com>